+++ src/interfaces/factive/index.ts
@@ -1,0 +1,3 @@
+export * from './IFactive';
+export * from './IFactiveCfg';
+export * from './IFactiveConstructor';