+++ src/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './factive/';
+export * from './api/';