+++ CHANGELOG.md
@@ -2,6 +2,51 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.7.0-alpha.2"></a>
+# [1.7.0-alpha.2](https://github.com/ArkadiumInc/node-inhabit-module-base/compare/v1.7.0-alpha.1...v1.7.0-alpha.2) (2019-04-19)
+
+
+
+<a name="1.7.0-alpha.1"></a>
+# [1.7.0-alpha.1](https://github.com/ArkadiumInc/node-inhabit-module-base/compare/v1.7.0-alpha.0...v1.7.0-alpha.1) (2019-04-19)
+
+
+
+<a name="1.7.0-alpha.0"></a>
+# [1.7.0-alpha.0](https://github.com/ArkadiumInc/node-inhabit-module-base/compare/v1.6.0...v1.7.0-alpha.0) (2019-04-19)
+
+
+### Features
+
+* **events:** Add playAgain and showBest events ([126bcae](https://github.com/ArkadiumInc/node-inhabit-module-base/commit/126bcae))
+
+
+
+<a name="1.6.0"></a>
+# [1.6.0](https://github.com/ArkadiumInc/node-inhabit-module-base/compare/v1.5.1...v1.6.0) (2017-08-31)
+
+
+### Features
+
+* **API:** Add more typings, add inhabitEvents ([914cd21](https://github.com/ArkadiumInc/node-inhabit-module-base/commit/914cd21))
+
+
+
+<a name="1.5.1"></a>
+## [1.5.1](https://github.com/ArkadiumInc/node-inhabit-module-base/compare/v1.5.0...v1.5.1) (2017-08-21)
+
+
+
+<a name="1.5.0"></a>
+# [1.5.0](https://github.com/ArkadiumInc/node-inhabit-module-base/compare/v1.5.0-alpha.3...v1.5.0) (2017-08-21)
+
+
+### Features
+
+* **TextClassificationService:** New version added (v2) ([27d275c](https://github.com/ArkadiumInc/node-inhabit-module-base/commit/27d275c))
+
+
+
 <a name="1.5.0-alpha.3"></a>
 # [1.5.0-alpha.3](https://github.com/ArkadiumInc/node-inhabit-module-base/compare/v1.5.0-alpha.2...v1.5.0-alpha.3) (2017-07-17)